+++ tests/test_typeclass/test_cache.py
@@ -14,7 +14,7 @@ class _MyABC(object, metaclass=ABCMeta):
         """Example abstract method."""
 
 
-class _MyConcete(_MyABC):
+class _MyConcrete(_MyABC):
     def get_number(self) -> int:
         """Concrete method."""
         return 1
@@ -34,18 +34,27 @@ def _my_abc(instance: _MyABC) -> int:
     return instance.get_number()
 
 
-def test_cache_invalidation():  # noqa: WPS218
+def test_cache_concrete(clear_cache) -> None:  # noqa: WPS218
     """Ensures that cache invalidation for ABC types work correctly."""
-    assert not my_typeclass._dispatch_cache  # noqa: WPS437
-    assert not my_typeclass._cache_token  # noqa: WPS437
+    with clear_cache(my_typeclass):
+        assert not my_typeclass._dispatch_cache  # noqa: WPS437
 
-    my_typeclass.instance(_MyABC)(_my_abc)
-    assert not my_typeclass._dispatch_cache  # noqa: WPS437
-    assert my_typeclass._cache_token  # noqa: WPS437
+        my_typeclass.instance(_MyABC)(_my_abc)
+        assert not my_typeclass._dispatch_cache  # noqa: WPS437
 
-    assert my_typeclass(_MyConcete()) == 1
-    assert my_typeclass._dispatch_cache  # noqa: WPS437
+        assert my_typeclass(_MyConcrete()) == 1
+        assert not my_typeclass._dispatch_cache  # noqa: WPS437
 
-    _MyABC.register(_MyRegistered)
-    assert my_typeclass(_MyRegistered()) == 2  # type: ignore
-    assert my_typeclass._dispatch_cache  # noqa: WPS437
+        _MyABC.register(_MyRegistered)
+        assert my_typeclass(_MyRegistered()) == 2  # type: ignore
+        assert not my_typeclass._dispatch_cache  # noqa: WPS437
+
+
+def test_cached_calls(clear_cache) -> None:
+    """Ensures that regular types trigger cache."""
+    with clear_cache(my_typeclass):
+        my_typeclass.instance(int)(_my_int)
+        assert not my_typeclass._dispatch_cache  # noqa: WPS437
+
+        assert my_typeclass(1)
+        assert my_typeclass._dispatch_cache  # noqa: WPS437