+++ tests/test_typeclass/test_callback.py
@@ -20,7 +20,7 @@ def _callback(
     return callback(instance)
 
 
-def test_callback():
+def test_callback() -> None:
     """Tests that callback works."""
     assert _callback('a', example) == 1
     assert _callback('abcd', example) == 4